replace custom cli code with cobra library
